@@ -1,10 +1,17 @@
 module nav_service
 
-go 1.15
+go 1.23.4
 
 require (
 	github.com/CatchZeng/dingtalk v1.2.0
 	github.com/robfig/cron/v3 v3.0.0
+	github.com/wwllss/zop v0.0.1
 	gorm.io/driver/mysql v1.1.2
 	gorm.io/gorm v1.22.0
+)
+
+require (
+	github.com/go-sql-driver/mysql v1.6.0 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.2 // indirect
 )
