@@ -1,12 +1,12 @@
 package main
 
 import (
+	"github.com/wwllss/zop"
 	"nav_service/hilog"
-	"nav_service/hop"
 	"time"
 )
 
-var RtTime hop.NavHandlerFunc = func(c *hop.Context) {
+var RtTime zop.NavHandlerFunc = func(c *zop.Context) {
 	now := time.Now()
 	c.Next()
 	hilog.Infof("[%d]-route:%s, RT:%v", c.StatusCode, c.Path, time.Since(now))
